requirements.txt
click==8.1.7
pyyaml==6.0.1
aiohttp==3.9.3
aiofiles==23.2.1
playwright==1.58.0
beautifulsoup4==4.12.3
# undetected-playwright==0.2.2  # Not available, using standard playwright with stealth
sqlalchemy==2.0.27
aiosqlite==0.19.0
tenacity==8.2.3
pydantic==2.6.1
python-dotenv==1.0.1
tqdm==4.66.2
requests==2.31.0
urllib3==2.1.0
pandas==2.2.1  # For reports and data analysis
